requirements.txt
Django>=4.2,<5.2
Pillow>=11.0.0
opencv-python-headless>=4.10.0
numpy>=2.0.0
psycopg2-binary>=2.9.9
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
